requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==10.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
